First batch of changes to permit VS2017 ARM64 tools to build required elements.

Related work items: #14268301

+++ ArmPkg/Drivers/ArmGic/GicV3/ArmGicV3Dxe.c
@@ -165,7 +165,7 @@ GicV3IrqInterruptHandler (
   UINT32                      GicInterrupt;
   HARDWARE_INTERRUPT_HANDLER  InterruptHandler;
 
-  GicInterrupt = ArmGicV3AcknowledgeInterrupt ();
+  GicInterrupt = (UINT32)ArmGicV3AcknowledgeInterrupt ();   // MS_CHANGE
 
   // Special Interrupts (ID1020-ID1023) have an Interrupt ID greater than the
   // number of interrupt (ie: Spurious interrupt).
@@ -389,8 +389,8 @@ GicV3DxeInitialize (
   // the system.
   ASSERT_PROTOCOL_ALREADY_INSTALLED (NULL, &gHardwareInterruptProtocolGuid);
 
-  mGicDistributorBase    = PcdGet64 (PcdGicDistributorBase);
-  mGicRedistributorsBase = PcdGet64 (PcdGicRedistributorsBase);
+  mGicDistributorBase    = (UINT32)PcdGet64 (PcdGicDistributorBase);           // MS_CHANGE
+  mGicRedistributorsBase = (UINT32)PcdGet64 (PcdGicRedistributorsBase);        // MS_CHANGE
   mGicNumInterrupts      = ArmGicGetMaxNumInterrupts (mGicDistributorBase);
 
   // We will be driving this GIC in native v3 mode, i.e., with Affinity
@@ -403,7 +403,7 @@ GicV3DxeInitialize (
     GicV3DisableInterruptSource (&gHardwareInterruptV3Protocol, Index);
 
     // Set Priority
-    RegOffset = Index / 4;
+    RegOffset = (UINT32)(Index / 4);   // MS_CHANGE
     RegShift = (Index % 4) * 8;
     MmioAndThenOr32 (
       mGicDistributorBase + ARM_GIC_ICDIPR + (4 * RegOffset),
@@ -432,7 +432,7 @@ GicV3DxeInitialize (
       for (Index = 8; Index < (mGicNumInterrupts / 4); Index++) {
         MmioWrite32 (
           mGicDistributorBase + ARM_GIC_ICDIPTR + (Index * 4),
-          CpuTarget
+          (UINT32)CpuTarget      // MS_CHANGE
           );
       }
     }
@@ -469,7 +469,7 @@ GicV3DxeInitialize (
     for (Index = 0; Index < (mGicNumInterrupts - 32); Index++) {
       MmioWrite32 (
         mGicDistributorBase + ARM_GICD_IROUTER + (Index * 8),
-        CpuTarget
+        (UINT32)CpuTarget
         );
     }
   }
